Toxic_content_classification_fullproj_week_2/requirements.txt
transformers==4.41.2
pillow==10.3.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
streamlit==1.40.2

